pyproject.toml
[tool.poetry]
name = "strawberry-graphene"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
strawberry-graphql = ">=0.87.2"
graphene = ">=3.0"
Django = ">=3.2.9"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.5b1"
flake8 = "^3.9.2"
flake8-black = "^0.2.1"
flake8-bugbear = "^21.4.3"
mypy = "^0.812"
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
